% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDF.R
\name{getDF}
\alias{getDF}
\title{Data Project: How could social media news influence the stock returns?}
\usage{
getDF(urls, key = Sys.getenv("NYTIMES_AS_KEY"))
}
\arguments{
\item{urls}{The URLs that generated from "makeURL" function}

\item{key}{Your New York Times API AS KEY}
}
\description{
The fluctuation in the stock market can be reflected and predicted through the most updated news.
This project takes news data as an illustration. Specifically, our goal is to find a method
to collect and combine both sturctured (numerical) data from Yahoo Finance and unsturctured (text)
data from NY Times together by web scraping data from APIs and further text cleaning.
}
\examples{
URLs <- makeURL(2000, 1, 2000, 2)
getDF(URLs, key = Sys.getenv("NYTIMES_AS_KEY"))
}
