% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeURL.R
\name{makeURL}
\alias{makeURL}
\title{Data Project: How could social media news influence the stock returns?}
\usage{
makeURL(begin_year, begin_month, end_year, end_month)
}
\arguments{
\item{begin_year}{The starting year of article data collections}

\item{begin_month}{The starting month of article data collections}

\item{end_year}{The ending year of article data collections}

\item{end_month}{The ending month of article data collections}
}
\description{
The fluctuation in the stock market can be reflected and predicted through the most updated news.
This project takes news data as an illustration. Specifically, our goal is to find a method
to collect and combine both sturctured (numerical) data from Yahoo Finance and unsturctured (text)
data from NY Times together by web scraping data from APIs and further text cleaning.
}
\examples{
makeURL(2000, 1, 2006, 3)
}
